support crc32 function

+++ gensrc/script/functions.py
@@ -326,7 +326,7 @@ vectorized_functions = [
 
     [30430, 'translate', 'VARCHAR', ['VARCHAR', 'VARCHAR', 'VARCHAR'], 'StringFunctions::translate',
      'StringFunctions::translate_prepare', 'StringFunctions::translate_close'],
-    [30431, 'crc32', 'INT', ['VARCHAR'], 'StringFunctions::crc32'],
+    [30431, 'crc32', 'BIGINT', ['VARCHAR'], 'StringFunctions::crc32'],
 
     # Binary Functions
     # to_binary
